copy: return 0 on success and appropriate errno on failure
This code was still using an old paradigm of returning the
number of bytes associated with a successful submission.
Just return 0 on success instead - if caller needs the
number of bytes for some reason they have the information
to get it.

While here, return an appropriate negated errno where possible -
we especially want ENOMEM returned when an ioat channel runs out
of descriptors.

+++ include/spdk_internal/copy_engine.h
@@ -45,9 +45,9 @@ struct spdk_copy_task {
 };
 
 struct spdk_copy_engine {
-	int64_t	(*copy)(void *cb_arg, struct spdk_io_channel *ch, void *dst, void *src,
+	int	(*copy)(void *cb_arg, struct spdk_io_channel *ch, void *dst, void *src,
 			uint64_t nbytes, spdk_copy_completion_cb cb);
-	int64_t	(*fill)(void *cb_arg, struct spdk_io_channel *ch, void *dst, uint8_t fill,
+	int	(*fill)(void *cb_arg, struct spdk_io_channel *ch, void *dst, uint8_t fill,
 			uint64_t nbytes, spdk_copy_completion_cb cb);
 	struct spdk_io_channel *(*get_io_channel)(void);
 };
